portfolio added to business

+++ FlexitHisBLL/Services/File/FileService.cs
@@ -1,23 +1,20 @@
 ﻿using Medicloud.BLL.DTO;
+using Medicloud.BLL.Services.FileUpload;
+using Medicloud.DAL.DAO;
 using Medicloud.DAL.Infrastructure.UnitOfWork;
 using Medicloud.DAL.Repository.File;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 namespace Medicloud.BLL.Services.File
 {
 	public class FileService:IFileService
 	{
 		private readonly IUnitOfWork _unitOfWork;
 		private readonly IFileRepository _fileRepository;
-
-		public FileService(IUnitOfWork unitOfWork, IFileRepository fileRepository)
+		private readonly IFileUploadService _fileUploadService;
+		public FileService(IUnitOfWork unitOfWork, IFileRepository fileRepository, IFileUploadService fileUploadService)
 		{
 			_unitOfWork = unitOfWork;
 			_fileRepository = fileRepository;
+			_fileUploadService = fileUploadService;
 		}
 
 		public async Task<FileDTO> GetById(int id)
@@ -36,7 +33,28 @@ namespace Medicloud.BLL.Services.File
 		public async Task DeleteById(int id)
 		{
 			using var con= _unitOfWork.BeginConnection();
-			await _fileRepository.DeleteFileAsync(id);
+			var file=await _fileRepository.GetFileByIdAsync(id);
+			if (file!=null)
+			{
+				bool isDeleted=_fileUploadService.DeleteFile(file.filePath);
+				if (isDeleted)
+				{
+					await _fileRepository.DeleteFileAsync(id);
+
+				}
+
+			}
+		}
+
+		public async Task<int> AddAsync(FileDTO dto)
+		{
+			var dao = new FileDAO
+			{
+				fileName = dto.fileName,
+				filePath = dto.filePath
+			};
+			using var con = _unitOfWork.BeginConnection();
+			return await _fileRepository.AddFileAsync(dao);
 		}
 	}
 }
